combine async with sync integration tests
